AI_Based_elearning-_recommendation_system-main/python-ml/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
pandas
sentence-transformers
python-dotenv
pydantic
pytest
httpx
